- renamed migration folder to database - moved Database Class into new database folder - status check for webserver is not protected anymore (externally check if application is running)
