+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mellow",
-  "version": "2.2.14",
+  "version": "2.2.17",
   "main": "src/index.js",
   "author": "void* <voidp@protonmail.com> (https://github.com/v0idp)",
   "repository": {
