+++ src/index.js
@@ -1,7 +1,7 @@
 const BotClient = require('./bots/DiscordBot.js');
 const WebServer = require('./web/WebServer.js');
-const Database = require('./Database.js');
-const { migrateALL } = require('./migration/migration.js');
+const Database = require('./database/Database.js');
+const { migrateALL } = require('./database/migration.js');
 const { version } = require('../package.json');
 
 console.log(`Mellow v${version}`);
